Safari-611.1.10/LayoutTests/fast/parser/style-in-create-contextual-fragment-expected.txt

| <html>
|   <head>
|     <script>
|       src="../../resources/dump-as-markup.js"
|     "\n"
|   <body>
|     "\n"
|     <script>
|       "\nvar range = document.createRange();\nrange.selectNode(document.body);\nvar documentFragment = range.createContextualFragment("<title>hi</title><style></style><body>XXX<div>ttt</div></body>");\ndocument.body.appendChild(documentFragment);\n"
|     <title>
|       "hi"
|     <style>
|     "XXX"
|     <div>
|       "ttt"
|     "\n\n"
